LayoutTests/platform/chromium/inspector/debugger/debugger-expand-scope-expected.txt


Test that sections representing scopes of the current call frame are expandable and contain correct data.

Debugger was enabled.
Evaluated script in console.
Script execution paused.

Dump scope sections:
Catch
    e: Error
Local
    innerFunctionLocalVar: 2012
    this: Window
    x: 2010
Closure
    makeClosureLocalVar: "local.TextParam"
    n: "TextParam"
WindowGlobal
    <section collapsed>
Script execution resumed.
Debugger was disabled.
